requirements.txt
Flask==3.0.0
supabase==2.10.0
python-dotenv==1.0.0
Werkzeug==3.0.1
reportlab==4.0.7
Pillow==10.1.0
python-dateutil==2.8.2
